Moved Helper.js and Validation.js to addition folder
