+++ screens/Confirm.js
@@ -1,6 +1,6 @@
 import { View, Text, Modal, Button } from "react-native";
 import React from "react";
-import { confirmStyles } from "../utils/Helper";
+import { confirmStyles } from "../addition/Helper";
 // this is a modal that shows the user data, user can choose to go back to sign up or continue to game
 
 export default function Confirm({
